socket-io-init: accept 'config.authorize' function
To auth socket.io connections.  Function should take a single argument,
which is the socket instance of the new connection.  It should
return a promise that resolves with a boolean value indicating whether
the connection is authorized.  If the promise resolves with a `false`
result or rejects, the socket is disconnected.

Fixes #66